fix: 리뷰 피드백 2차 반영

+++ src/main/java/com/nhnacademy/ssacthree_shop_api/orderset/deliveryrule/dto/DeliveryRuleGetResponse.java
@@ -11,8 +11,8 @@ public class DeliveryRuleGetResponse {
 
     private long deliveryRuleId;
     private String deliveryRuleName;
-    private long deliveryFee;
-    private long deliveryDiscountCost;
+    private int deliveryFee;
+    private int deliveryDiscountCost;
     private boolean deliveryRuleIsSelected;
     private LocalDateTime deliveryRuleCreatedAt;
 }
